+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * json: deprecate KeyConverter and ValueConverter in favor of the ones in `klite-core` module
 * server: basicAuth() helper
 * jdbc: `DockerCompose.startDB()` will now wait until DB actually accepts connections
+* jdbc: support kotlin.uuid.Uuid and kotlin.time.Instant by default
 
 # 1.7.2
 * server: HttpServer now accepts connections inside of workerPool, not single thread - for better scalability under very heavy load